update : puppy API response 값 수정

+++ nolmung-be/src/main/java/com/ssafy/nolmung/walk/domain/Walk.java
@@ -23,7 +23,7 @@ public class Walk {
     private int walkId;
 
     @Column(name = "walk_distance")
-    private int walkDistance;
+    private double walkDistance;
 
     @Column(name = "walk_start_time")
     private LocalDateTime walkStartTime;
